# perf: use faster walkdir for file walking