Walk is less efficient than WalkDir, introduced in Go 1.16,
which avoids calling os.Lstat on every visited file or directory.

See https://go.dev/doc/go1.16#path/filepath

Benchmark results:

```
name        old time/op    new time/op    delta
Run/dot-16    1.05ms ± 8%    0.89ms ± 7%  -15.43%  (p=0.000 n=10+10)

name        old alloc/op   new alloc/op   delta
Run/dot-16     241kB ± 0%     225kB ± 0%   -6.94%  (p=0.000 n=10+10)

name        old allocs/op  new allocs/op  delta
Run/dot-16     2.49k ± 0%     2.37k ± 0%   -4.62%  (p=0.000 n=10+10)
```